requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==1.4.1
MarkupSafe==1.1.1
numpy==1.18.5
opencv-python==4.2.0.34
Pillow==7.1.2
requests==2.23.0
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
